2 nd commit

+++ social/src/main/java/com/example/social/response/JwtResponse.java
@@ -3,16 +3,15 @@ package com.example.social.response;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private Long id;
+    private int id;
     private String username;
     private String email;
 
-    public JwtResponse(String accessToken, Long id, String username, String email) {
+    public JwtResponse(String accessToken, int id, String username, String email) {
         this.token = accessToken;
         this.id = id;
         this.username = username;
         this.email = email;
-
     }
 
     public String getAccessToken() {
@@ -31,11 +30,11 @@ public class JwtResponse {
         this.type = tokenType;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
